Rename music reducer
