+++ src/presentation/redux/getMusicInfo/reducer.ts
