+++ src/presentation/routes/home/components/musicList/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { MusicPanel } from './music';
 import { Dispatch, bindActionCreators } from 'redux';
-import { requestSebaCurationMusicLists, RequestSebaMusic } from 'presentation/redux/music/actions';
+import { requestSebaCurationMusicLists, RequestSebaMusic } from 'presentation/redux/getMusicInfo/actions';
 import { Music } from 'domain/entity/music';
-import { SebaMusicStateProps } from 'presentation/redux/music/reducer';
+import { SebaMusicStateProps } from 'presentation/redux/getMusicInfo/reducer';
 import { FetchResult } from 'presentation/redux/actionTypes';
 import { FetchStatus } from 'presentation/redux/FetchStatus';
 import { Loading } from 'presentation/components/loading';
